requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
click==7.1.2              # via pip-tools
colorzero==1.1            # via gpiozero
gpiozero==1.5.1           # via -r requirements.in
json-minify==0.3.0        # via -r requirements.in
pathlib==1.0.1            # via -r requirements.in
pip-tools==5.1.1          # via -r requirements.in
pyserial==3.4             # via -r requirements.in
six==1.14.0               # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
